Fixed install script.

+++ install.sh
@@ -1,45 +1,41 @@
 #!/bin/bash
-if [ -e $HOME/.vimrc ]
+if [ -e $HOME/.vimrc ] && [ ! -L $HOME/.vimrc ]
 then
-   if [ -h $HOME/.vimrc ]
-   then
-      rm $HOME/.vimrc
-      echo "Removed symbolic link for .vimrc."
-   else
-      echo "vimrc not a symlink."
-      mv $HOME/.vimrc $HOME/.vimrc.old
-   fi
+   mv $HOME/.vimrc $HOME/.vimrc.old
 fi
-ln -s ./vimrc $HOME/.vimrc
+ln -sf $HOME/.dotfiles/vimrc $HOME/.vimrc
 
-if [ -d $HOME/.vim ]
+if [ -d $HOME/.vim ] && [ ! -L $HOME/.vim ]
 then
    mv $HOME/.vim $HOME/.vim.old
 fi
-ln -s ./vim $HOME/.vim
+ln -sf $HOME/.dotfiles/vim $HOME/.vim
 
-if [ -e $HOME/.zshrc ]
+if [ -e $HOME/.zshrc ] && [ ! -L $HOME/.zshrc ]
 then
    mv $HOME/.zshrc $HOME/.zshrc.old
 fi
-ln -s ./zshrc $HOME/.zshrc
+ln -sf $HOME/.dotfiles/zshrc $HOME/.zshrc
 
-if [ -e $HOME/.gitconfig ]
+if [ -e $HOME/.gitconfig ] && [ ! -L $HOME/.gitconfig ]
 then
    mv $HOME/.gitconfig $HOME/.gitconfig.old
 fi
-ln -s ./gitconfig $HOME/.gitconfig
+ln -sf $HOME/.dotfiles/gitconfig $HOME/.gitconfig
 
-if [ -e $HOME/.Xresources ]
+if [ -e $HOME/.Xresources ] && [ ! -L $HOME/.Xresources ]
 then
    mv $HOME/.Xresources $HOME/.Xresources.old
 fi
-ln -s ./Xresources $HOME/.Xresources
+ln -sf $HOME/.dotfiles/Xresources $HOME/.Xresources
 
 if [ -d $HOME/.oh-my-zsh ]
 then
-   mv $HOME/.oh-my-zsh/themes/mine.zsh-theme $HOME/.oh-my-zsh/themes/mine.zsh-theme.old
-   ln -s ./mine.zsh-theme $HOME/.oh-my-zsh/themes/mine.zsh-theme
+   if [ ! -L $HOME/.oh-my-zsh/themes/mine.zsh-theme ]
+   then
+      mv $HOME/.oh-my-zsh/themes/mine.zsh-theme $HOME/.oh-my-zsh/themes/mine.zsh-theme.old
+   fi
+   ln -sf $HOME/.dotfiles/mine.zsh-theme $HOME/.oh-my-zsh/themes/mine.zsh-theme
 else
    echo "Could not install mine.zsh-theme. Try installing oh-my-zsh if it isn't already installed."
 fi
